hl-line: Box the current line instead of underline + bold
The box is easier to see, and underline + bold weren't all that good. Especially
the bold... made it harder to see the current line when there were bold lines
surrounding it. Also made it harder to see certain things (like in Elfeed, if
the current item is read or unread - since hl-line's bold obstructed the unread
bold).

Signed-off-by: Gergely Nagy <emacs@gergo.csillger.hu>